+++ src/Empathy/MVC/Empathy.php
@@ -2,7 +2,7 @@
 
 namespace Empathy\MVC;
 
-define('MVC_VERSION', '1.4.1');
+define('MVC_VERSION', '2.0.0');
 
 
 /**