+++ src/Empathy/MVC/Entity.php
@@ -5,20 +5,8 @@ namespace Empathy\MVC;
 use Empathy\MVC\Config;
 use Empathy\MVC\LogItem;
 
-/**
- * Empathy Entity
- * @package         Empathy
- * @file            Empathy/Entity.php
- * @description     Simple "ORM style" model objects for Empathy.
- * @author          Mike Whiting
- * @license         See LICENCE
- *
- * (c) copyright Mike Whiting
- * with this source code in the file licence.txt
- */
 class Entity
 {
-
     public function __construct()
     {
         if (Config::get('LEGACY_MODEL')) {
@@ -26,46 +14,12 @@ class Entity
         }
     }
 
-
-    /**
-     * The name of the database table the entity maps to.
-     */
     const TABLE = '';
 
-    /**
-     * Sanitize constant
-     */
-    const SANITIZE = 1;
-
-    /**
-     * Sanitize no post constant
-     */
-    const SANITIZE_NO_POST = 2;
-
-    /**
-     * Validation object for 'model'.
-     */
+    private const GLOBALLY_IGNORED_PROPERTIES = ['id', 'table']; // leaving in table to support old models
     private $val;
-
-    /**
-     * The last query results.
-     */
     private $result;
-
-    /**
-     * Fields to ignore while creating/updating records.
-     *
-     */
-    private $globally_ignored_property = array('id', 'table'); // leaving in table to support old models
-
-    /**
-     * The current model properties/fields observed through reflrection.
-     */
     private $properties;
-
-    /**
-     * The PDO database connection handle.
-     */
     private $dbh;
 
     /**
@@ -73,47 +27,23 @@ class Entity
      */
     protected static $table = '';
 
-    /**
-     * Get the current time in the form of a MySQL-friendly time and date stamp.
-     *
-     * @return string date stamp
-     */
     public function MYSQLTime()
     {
         return '\'' . date('Y:m:d H:i:s', time()) . '\'';
     }
 
-
-    /**
-     * Instantiates validation object
-     * and loads model properties/fields.
-     *
-     * @return void
-     */
     public function init()
     {
         $this->val = new Validate();
-        $this->properties = array();
+        $this->properties = [];
         $this->loadProperties();
     }
 
-    /**
-     * Gets the last auto-incremented ID from MySQL.
-     *
-     * @return integer $id
-     */
     public function insertId()
     {
         return $this->dbh->lastInsertId();
     }
 
-    /**
-     * Loads the properties/fields of the current mode.
-     * Get properties from all subclasses in the right order
-     * (with a bit of hackery to make 'table' last)
-     *
-     * @return void
-     */
     private function loadProperties()
     {
         $super_class = 'Empathy\MVC\Entity';
@@ -121,13 +51,13 @@ class Entity
         $r = new \ReflectionClass(get_class($this));
 
         if ($r->getParentClass()->getName() != $super_class) {
-            $props = array();
+            $props = [];
             while (($class = $r->getName()) != $super_class) {
                 $props[] = $r->getProperties();
                 $r = $r->getParentClass();
             }
             $props = array_reverse($props);
-            $properties = array();
+            $properties = [];
             foreach ($props as $p) {
                 foreach ($p as $rp) {
                     $name = $rp->name;
@@ -146,14 +76,12 @@ class Entity
                 }
             }
         }
-        $this->properties = array_diff($this->properties, $this->globally_ignored_property);
+        $this->properties = array_diff($this->properties, self::GLOBALLY_IGNORED_PROPERTIES);
     }
 
     /**
-     * Connect to database. Only used in old fashoined applications
+     * Used in old fashioned applications
      * where Entities are not loaded through the Model class.
-     *
-     * @return @void
      */
     public function dbConnect()
     {
@@ -166,11 +94,6 @@ class Entity
         if (!defined('DB_USER') || DB_USER == '') {
             throw new SafeException('DB Error: No database username');
         }
-        /* commenting this out because it's too annoying
-           if (!defined('DB_PASS') || DB_PASS == '') {
-           throw new SafeException('DB Error: No database password');
-           }
-        */
         $dsn = 'mysql:host=' . DB_SERVER . ';dbname=' . DB_NAME . ';';
         if (defined('DB_PORT') && is_numeric(DB_PORT)) {
             $dsn .= 'port=' . DB_PORT . ';';
@@ -178,24 +101,11 @@ class Entity
         $this->dbh = new \PDO($dsn, DB_USER, DB_PASS);
     }
 
-    /**
-     * Set the model objects database connecion handler
-     *
-     * @param PDO Handle $dbh
-     *
-     * @return void
-     */
     public function setDBH(&$dbh)
     {
         $this->dbh = $dbh;
     }
 
-
-    /**
-     * Clear associated PDO objects
-     *
-     * @return void
-     */
     public function dbDisconnect()
     {
         unset($this->result);
@@ -206,9 +116,9 @@ class Entity
     {
         $log = new LogItem(
             'sql query',
-            array(
+            [
                 'query' => $sql
-            ),
+            ],
             self::class,
             $level
         );
@@ -224,22 +134,13 @@ class Entity
         $log->fire();
     }
 
-    /**
-     * Perform MySQL query
-     *
-     * @param string $sql the SQL query
-     *
-     * @param string $error the error to product on failure
-     *
-     * @return void
-     */
-    public function query($sql, $error = '', $params = array())
+    public function query($sql, $error = '', $params = [])
     {
         $result = null;
-        $errors = array('', '', '');
+        $errors = ['', '', ''];
         try {
             if (sizeof($params)) {
-                $sth = $this->dbh->prepare($sql, array(\PDO::ATTR_ERRMODE => \PDO::ERRMODE_EXCEPTION));
+                $sth = $this->dbh->prepare($sql, [\PDO::ATTR_ERRMODE => \PDO::ERRMODE_EXCEPTION]);
                 $sth->execute($params);
                 $result = $sth;
             } else {
@@ -258,24 +159,16 @@ class Entity
             $this->logQuery($sql, $error . $errors[2], $params, 'error');
             throw $e;
         }
-        
         return $result;
     }
 
-    /**
-     * Load a record from a table based on the current id
-     *
-     * @param string $table Older applications assume they need to pass their table names but this was worked around
-     *
-     * @return boolean success of load
-     */
-    public function load($table = null)
+    public function load($id = 0)
     {
-        if ($table == null) {
-            $c = get_class($this);
-            $table = $c::TABLE;
+        if ($id > 0) {
+            $this->id = $id;
         }
 
+        $table = $this::TABLE;
         $loaded = true;
         $sql = "SELECT * FROM $table WHERE id = $this->id";
         $error = "Could not load record from $table.";
@@ -291,32 +184,20 @@ class Entity
         } else {
             $loaded = false;
         }
-
         return $loaded;
     }
 
-    /**
-     * Loads one column from all the rows from in a table and returns a data structure that
-     * uses the row id as an index. Useful for building html select form elemements (hence the name).
-     *
-     * @param string $table the table to load data from
-     *
-     * @param string field the column to load
-     *
-     * @param string order the field to order the results by
-     *
-     * @return array data structure of 'options'
-     */
-    public function loadAsOptions($table, $field, $order = null)
+    public function loadAsOptions($field, $order = null)
     {
-        $data = array();
-        $sql = 'SELECT id,' . $field . ' FROM ' . $table;
+        $data = [];
+        $table = $this::TABLE;
+        $sql = "SELECT id, $field FROM $table";
         if ($order !== null && $order != '') {
-            $sql .= ' ORDER BY ' . $order;
+            $sql .= " ORDER BY $order";
         } else {
-            $sql .= ' ORDER BY ' . $field;
+            $sql .= " ORDER BY $field";
         }
-        $error = 'Could not load ' . $table . ' as options';;
+        $error = "Could not load $table as options";
         $result = $this->query($sql, $error);
         foreach ($result as $row) {
             $id = $row['id'];
@@ -326,173 +207,76 @@ class Entity
         return $data;
     }
 
-    /**
-     * Escape non-numeric and non-empty fields
-     * Check the field exists within the POST data by default
-     *
-     * @return void
-     */
-    public function sanitize($checkPostValues = true)
+    public function save($filter = [])
     {
-        foreach ($this->properties as $property) {
-            if ((!$checkPostValues || ($checkPostValues && isset($_POST[$property]))) &&
-                !in_array($property, $this->globally_ignored_property) &&
-                $this->$property !== null &&
-                !is_numeric($this->$property)) {
-                $this->$property = substr($this->dbh->quote($this->$property), 1, -1);
-            }
-        }
-    }
-
-
-    /**
-     * Escape non-numeric and non-empty fields.
-     *
-     * @return void
-     */
-    public function sanitizeNoPost()
-    {
-        $this->sanitize(false);
-    }
-
-
-    /**
-     * Save object back to database (update query).
-     *
-     * @param string $table the table to save to.
-     *
-     * @param array $format List of fields to apply HTML filtering to.
-     *
-     * @param integer $sanitize the form of anti-injection attack sanitization to apply.
-     * Either none, while checking POST data, while not checking POST data.
-     *
-     * @return void
-     */
-    public function save($table = null, $format = null, $sanitize = self::SANITIZE_NO_POST)
-    {
-        if ($table === null) {
-            $table = $this::TABLE;
-        }
-
-        if ($format === null) {
-            $format = array();
-        }
-
-        $this->toXHTMLChris($format);
-        if ($sanitize == self::SANITIZE) {
-            $this->sanitize();
-        } elseif ($sanitize == self::SANITIZE_NO_POST) {
-            $this->sanitizeNoPost();
-        }
-
+        $this->toFilteredHTML($filter);
+        $table = $this::TABLE;
         $sql = "UPDATE $table SET ";
 
-        $properties = array();
-
-        foreach ($this->properties as $property) {
-            array_push($properties, $property);
-        }
-
         $i = 0;
-        foreach ($properties as $property) {
+        $params = [];
+        foreach ($this->properties as $property) {
             $sql .= "$property = ";
-            if (is_numeric($this->$property) && !is_string($this->$property)) {
-                $sql .= $this->$property;
-            } elseif ($this->$property == '') {
+
+            if ($this->$property == '') {
                 $sql .= 'NULL';
             } elseif ($this->$property == 'DEFAULT') {
                 $sql .= 'DEFAULT';
             } elseif ($this->$property == 'MYSQLTIME') {
                 $sql .= $this->MYSQLTime();
             } else {
-                $sql .= "'" . $this->$property . "'";
+                $sql .= '?';
+                $params[] = $this->$property;
             }
-
-            if ($i + 1 < sizeof($properties)) {
-                $sql .= ", ";
+            if ($i + 1 < sizeof($this->properties)) {
+                $sql .= ', ';
             }
             $i++;
         }
         $sql .= " WHERE id = $this->id";
-
         $error = "Could not update table '$table'";
-
-        $this->query($sql, $error);
+        $this->query($sql, $error, $params);
     }
 
-    /**
-     * Insert object into database.
-     *
-     * @param string $table the table to save to.
-     *
-     * @param boolean $id Whether or not the table has an auto-increment id field.
-     *
-     * @param array $format List of fields to apply HTML filtering to.
-     *
-     * @param integer $sanitize the method of sanitization for non-numeric fields.
-     * Either none, while checking POST data, while not checking POST data.
-     *
-     * @return void
-     */
-    public function insert($table, $id, $format, $sanitize, $force_id = false)
+    public function insert($filter = [], $id = true)
     {
-        $this->toXHTMLChris($format);
-        if ($sanitize == self::SANITIZE) {
-            $this->sanitize();
-        } elseif ($sanitize == self::SANITIZE_NO_POST) {
-            $this->sanitizeNoPost();
-        }
-
-        $sql = 'INSERT INTO ' . $table . ' VALUES(';
+        $this->toFilteredHTML($filter);
+        $table = $this::TABLE;
+        $sql = "INSERT INTO $table VALUES(";
         if ($id) {
             $sql .= 'NULL, ';
         }
-
         $i = 0;
-        $id = 0;
-
+        $params = [];
         foreach ($this->properties as $property) {
-            if (($force_id && $property == 'id') || !$force_id) {
-                if (is_numeric($this->$property) && !is_string($this->$property)) {
-                    $sql .= $this->$property;
-                } elseif ($this->$property == '') {
-                    $sql .= 'NULL';
-                } elseif ($this->$property == 'DEFAULT') {
-                    $sql .= 'DEFAULT';
-                } elseif ($this->$property == 'MYSQLTIME') {
-                    $sql .= $this->MYSQLTime();
-                } else {
-                    $sql .= "'" . $this->$property . "'";
-                }
-
-                if (($i + 1) < sizeof($this->properties)) {
-                    $sql .= ", ";
-                }
+            if ($this->$property == '') {
+                $sql .= 'NULL';
+            } elseif ($this->$property == 'DEFAULT') {
+                $sql .= 'DEFAULT';
+            } elseif ($this->$property == 'MYSQLTIME') {
+                $sql .= $this->MYSQLTime();
+            } else {
+                $sql .= '?';
+                $params[] = $this->$property;
+            }
+            if (($i + 1) < sizeof($this->properties)) {
+                $sql .= ', ';
             }
             $i++;
         }
-        $sql .= ")";
+        $sql .= ')';
         $error = "Could not insert to table '$table'";
-        $this->query($sql, $error);
+        $this->query($sql, $error, $params);
 
         return $this->insertId();
     }
 
-    /**
-     * Gets all rows from a table and returns array.*
-     * @param string $table the table to fetch from.
-     *
-     * @return array $all all rows from the table.
-     */
-    public function getAll($table = null)
+    public function getAll()
     {
-        if ($table === null) {
-            $table = $this::TABLE;
-        }
-
-        $all = array();
-        $sql = 'SELECT * FROM ' . $table;
-        $error = 'Could not get all rows from ' . $table;
+        $table = $this::TABLE;
+        $all = [];
+        $sql = "SELECT * FROM $table";
+        $error = "Could not get all rows from '$table'";
         $result = $this->query($sql, $error);
 
         $i = 0;
@@ -504,20 +288,12 @@ class Entity
         return $all;
     }
 
-    /**
-     * Similar to getAll but with custom SQL appended to the query.
-     *
-     * @param string $table the target table.
-     *
-     * @param string $sql_string the additional custom SQL.
-     *
-     * @return array $all the rows returned as a result of the query
-     */
-    public function getAllCustom($table, $sql_string, $params = array())
+    public function getAllCustom($sqlString, $params = [])
     {
-        $all = array();
-        $sql = 'SELECT * FROM ' . $table . ' ' . $sql_string;
-        $error = 'Could not get all rows from ' . $table;
+        $table = $this::TABLE;
+        $all = [];
+        $sql = "SELECT * FROM $table $sqlString";
+        $error = "Could not get all rows from '$table'";
         $result = $this->query($sql, $error, $params);
 
         $i = 0;
@@ -525,20 +301,19 @@ class Entity
             $all[$i] = $row;
             $i++;
         }
-
         return $all;
     }
 
-    public function getPaginatePages($table, $sql_string, $page, $per_page)
+    public function getPaginatePages($sqlString, $page, $perPage, $params = [])
     {
-        $nav = array();
-        $sql = 'SELECT * FROM ' . $table . ' ' . $sql_string;
+        $table = $this::TABLE;
+        $nav = [];
+        $sql = "SELECT * FROM $table $sqlString";
         //$sql = 'SELECT FOUND_ROWS()';
-        $error = 'Could not get rows from ' . $table;
-        $result = $this->query($sql, $error);
+        $error = "Could not get rows from '$table'";
+        $result = $this->query($sql, $error, $params);
         $rows = $result->rowCount();
-        $p_rows = $rows;
-        $pages = ceil($rows / $per_page);
+        $pages = ceil($rows / $perPage);
         $i = 1;
         while ($i <= $pages) {
             if ($i == $page) {
@@ -548,26 +323,23 @@ class Entity
             }
             $i++;
         }
-
         return $nav;
     }
 
-    public function getPaginatePagesSimpleJoin($select, $table1, $table2, $sql_string, $page, $per_page, $leftJoins = '')
+    public function getPaginatePagesSimpleJoin($select, $table2, $sqlString, $page, $perPage, $leftJoins = '', $params = [])
     {
-        $nav = array();
-        $sql = 'SELECT ' . $select . ' FROM ' . $table1 . ' t1';
+        $nav = [];
+        $table1 = $this::TABLE;
+        $sql = "SELECT $select FROM $table1 t1";
 
         if ($leftJoins === '') {
             $sql .= ', ';
         }
-        $sql .= $leftJoins . $table2 . ' t2 ' . $sql_string;
-
-        $error = 'Could not get rows from ' . $table1;
-
-        $result = $this->query($sql, $error);
+        $sql .= "$leftJoins $table2 t2 $sqlString";
+        $error = "Could not get rows from '$table1'";
+        $result = $this->query($sql, $error, $params);
         $rows = $result->rowCount();
-        $p_rows = $rows;
-        $pages = ceil($rows / $per_page);
+        $pages = ceil($rows / $perPage);
         $i = 1;
         while ($i <= $pages) {
             if ($i == $page) {
@@ -577,19 +349,18 @@ class Entity
             }
             $i++;
         }
-
         return $nav;
     }
 
-    public function getPaginatePagesMultiJoin($select, $table1, $table2, $table3, $sql_string, $page, $per_page)
+    public function getPaginatePagesMultiJoin($select, $table2, $table3, $sqlString, $page, $perPage, $params = [])
     {
-        $nav = array();
-        $sql = 'SELECT ' . $select . ' FROM ' . $table1 . ' t1, ' . $table2 . ' t2, ' . $table3 . ' t3 ' . $sql_string;
-        $error = 'Could not get rows from ' . $table1;
-        $result = $this->query($sql, $error);
+        $nav = [];
+        $table1 = $this::TABLE;
+        $sql = "SELECT $select FROM $table1 t1, $table2 t2, $table3 t3 $sqlString";
+        $error = "Could not get rows from '$table1'";
+        $result = $this->query($sql, $error, $params);
         $rows = $result->rowCount();
-        $p_rows = $rows;
-        $pages = ceil($rows / $per_page);
+        $pages = ceil($rows / $perPage);
         $i = 1;
         while ($i <= $pages) {
             if ($i == $page) {
@@ -599,30 +370,28 @@ class Entity
             }
             $i++;
         }
-
         return $nav;
     }
 
     public function getPaginatePagesMultiJoinGroup(
         $select,
-        $table1,
         $table2,
         $table3,
-        $sql_string,
+        $sqlString,
         $page,
-        $per_page,
+        $perPage,
         $group,
-        $order
-    )
-    {
-        $nav = array();
-        $sql = 'SELECT ' . $select . ' FROM ' . $table1 . ' t1, ' . $table2 . ' t2, ' . $table3 . ' t3 ' . $sql_string;
-        $sql .= ' GROUP BY ' . $group . ' ORDER BY ' . $order;
-        $error = 'Could not get rows from ' . $table1;
-        $result = $this->query($sql, $error);
+        $order,
+        $params = []
+    ) {
+        $nav = [];
+        $table1 = $this::TABLE;
+        $sql = "SELECT $select FROM $table1 t1, $table2 t2, $table3 t3 $sqlString";
+        $sql .= " GROUP BY $group ORDER BY $order";
+        $error = "Could not get rows from '$table1'";
+        $result = $this->query($sql, $error, $params);
         $rows = $result->rowCount();
-        $p_rows = $rows;
-        $pages = ceil($rows / $per_page);
+        $pages = ceil($rows / $perPage);
         $i = 1;
         while ($i <= $pages) {
             if ($i == $page) {
@@ -632,92 +401,89 @@ class Entity
             }
             $i++;
         }
-
         return $nav;
     }
 
-    public function getAllCustomPaginate($table, $sql_string, $page, $per_page)
+    public function getAllCustomPaginate($sqlString, $page, $perPage, $params = [])
     {
-        $all = array();
-        $start = ($page - 1) * $per_page;
-        $sql = 'SELECT * FROM ' . $table . ' ' . $sql_string . ' LIMIT ' . $start . ', ' . $per_page;
-        //$sql = 'SELECT SQL_CALC_FOUND_ROWS * FROM '.$table.' '.$sql_string.' LIMIT '.$start.', '.$per_page;
-        $error = 'Could not get rows from ' . $table;
-
-        $result = $this->query($sql, $error);
+        $all = [];
+        $table = $this::TABLE;
+        $start = ($page - 1) * $perPage;
+        $sql = "SELECT * FROM $table $sqlString LIMIT $start, $perPage";
+        $error = "Could not get rows from '$table'";
+        $result = $this->query($sql, $error, $params);
         $i = 0;
         foreach ($result as $row) {
             $all[$i] = $row;
             $i++;
         }
-
         return $all;
     }
 
-    public function getAllCustomPaginateSimpleJoin($select, $table1, $table2, $sql_string, $page, $per_page, $leftJoins = '')
+    public function getAllCustomPaginateSimpleJoin($select, $table2, $sqlString, $page, $perPage, $leftJoins = '', $params = [])
     {
-        $all = array();
-        $start = ($page - 1) * $per_page;
-        $sql = 'SELECT ' . $select . ' FROM ' . $table1 . ' t1';
+        $all = [];
+        $table1 = $this::TABLE;
+        $start = ($page - 1) * $perPage;
+        $sql = "SELECT $select FROM $table1 t1";
 
         if ($leftJoins === '') {
             $sql .= ', ';
         }
-        $sql .= $leftJoins . $table2 . ' t2 ' . $sql_string . ' LIMIT ' . $start . ', ' . $per_page;
-        $error = 'Could not get rows from ' . $table1;
+        $sql .= "$leftJoins $table2 t2 $sqlString LIMIT $start, $perPage";
+        $error = "Could not get rows from '$table1'";
 
-        $result = $this->query($sql, $error);
+        $result = $this->query($sql, $error, $params);
         $i = 0;
         foreach ($result as $row) {
             $all[$i] = $row;
             $i++;
         }
-
         return $all;
     }
 
-    public function getAllCustomPaginateMultiJoin($select, $table1, $table2, $table3, $sql_string, $page, $per_page)
+    public function getAllCustomPaginateMultiJoin($select, $table2, $table3, $sqlString, $page, $perPage, $params = [])
     {
-        $all = array();
-        $start = ($page - 1) * $per_page;
-        $sql = 'SELECT ' . $select . ' FROM ' . $table1 . ' t1, ' . $table2 . ' t2, '
-            . $table3 . ' t3 ' . $sql_string . ' LIMIT ' . $start . ', ' . $per_page;
-        $error = 'Could not get rows from ' . $table1;
-        $result = $this->query($sql, $error);
+        $all = [];
+        $table1 = $this::TABLE;
+        $start = ($page - 1) * $perPage;
+        $sql = "SELECT $select FROM $table1 t1, $table2 t2,"
+            . " $table3 t3 $sqlString LIMIT $start, $perPage";
+        $error = "Could not get rows from '$table1'";
+        $result = $this->query($sql, $error, $params);
         $i = 0;
         foreach ($result as $row) {
             $all[$i] = $row;
             $i++;
         }
-
         return $all;
     }
 
     public function getAllCustomPaginateMultiJoinGroup(
         $select,
-        $table1,
         $table2,
         $table3,
-        $sql_string,
+        $sqlString,
         $page,
-        $per_page,
+        $perPage,
         $group,
-        $order
+        $order,
+        $params = []
     )
     {
-        $all = array();
-        $start = ($page - 1) * $per_page;
-        $sql = 'SELECT ' . $select . ' FROM ' . $table1 . ' t1, ' . $table2 . ' t2, '
-            . $table3 . ' t3 ' . $sql_string . ' GROUP BY ' . $group
-            . ' ORDER BY ' . $order . ' LIMIT ' . $start . ', ' . $per_page;
-        $error = 'Could not get rows from ' . $table1;
-        $result = $this->query($sql, $error);
+        $all = [];
+        $table1 = $this::TABLE;
+        $start = ($page - 1) * $perPage;
+        $sql = "SELECT $select FROM $table1 t1, $table2 t2,"
+            . " $table3 t3 $sqlString GROUP BY $group"
+            . " ORDER BY $order LIMIT $start, $perPage";
+        $error = "Could not get rows from '$table1'";
+        $result = $this->query($sql, $error, $params);
         $i = 0;
         foreach ($result as $row) {
             $all[$i] = $row;
             $i++;
         }
-
         return $all;
     }
 
@@ -725,18 +491,17 @@ class Entity
     {
         foreach ($this->properties as $property) {
             if (($force_id && $property == 'id') ||
-                (!in_array($property, $this->globally_ignored_property)
+                (!in_array($property, self::GLOBALLY_IGNORED_PROPERTIES)
                     && !in_array($property, $ignore))) {
                 $this->$property = $_POST[$property];
             }
         }
     }
 
-
-    public function prepareOptions($first, $label, $table)
+    public function prepareOptions($first, $label)
     {
-        $option = array();
-        $data = $this->getAll($table);
+        $option = [];
+        $data = $this->getAll();
         if ($first != '') {
             $option[0] = $first;
         }
@@ -749,9 +514,7 @@ class Entity
         return $option;
     }
 
-    // @todo deprecate function name?
-    // `toFilteredHTML` might be better?
-    public function toXHTMLChris($formatting)
+    public function toFilteredHTML($filtering)
     {
         $pTagPattern = '!&lt;p&gt;(.*?)&lt;/p&gt;!m';
         $aTagPattern = '!&lt;a +href=&quot;((?:ht|f)tps?://.*?)&quot;'
@@ -764,7 +527,7 @@ class Entity
         $preTagPattern2 = '!&lt;pre(?: +class=&quot;(.*?)&quot;)? *&gt;\n*(.*?)&lt;/pre&gt;!ms';
 
         foreach ($this->properties as $property) {
-            if (!is_numeric($property) && in_array($property, $formatting)) {
+            if (!is_numeric($property) && in_array($property, $filtering)) {
                 $markup = $this->$property;
                 $markup = str_replace("\r", "\n", $markup);
                 $markup = preg_replace("!\n\n+!", "\n", $markup);
@@ -829,86 +592,18 @@ class Entity
         }
     }
 
-    public function toXHTML($formatting)
-    {
-        foreach ($this->properties as $property) {
-            if (!(is_numeric($property))) {
-                //$this->$property = ereg_replace(38, "&amp;", $this->$property);
-
-                if (in_array($property, $formatting)) {
-                    $this->$property = str_replace("\r\n", '<br />', $this->$property);
-                } else {
-                    $this->$property = str_replace("\r\n", ' ', $this->$property);
-                }
-            }
-        }
-    }
-
     public function buildUnionString($ids)
     {
         array_unshift($ids, 0);
         return '(' . implode(',', array_unique($ids)) . ')';
     }
 
-    public function stripMSWordChars()
+    public function delete()
     {
-        foreach ($this->properties as $property) {
-            if (!(is_numeric($property))) {
-                /*
-                  $this->$property = ereg_replace(133, "&#133;", $this->$property); // ellipses
-                  $this->$property = ereg_replace(8226, "&#8243;", $this->$property); // double prime
-                  $this->$property = ereg_replace(8216, "&#039;", $this->$property); // left single quote
-                  $this->$property = ereg_replace(145, "&#039;", $this->$property); // left single quote
-                  $this->$property = ereg_replace(8217, "&#039;", $this->$property); // right single quote
-                  $this->$property = ereg_replace(146, "&#039;", $this->$property); // right single quote
-                  $this->$property = ereg_replace(8220, "&#034;", $this->$property); // left double quote
-                  $this->$property = ereg_replace(147, "&#034;", $this->$property); // left double quote
-                  $this->$property = ereg_replace(8221, "&#034;", $this->$property); // right double quote
-                  $this->$property = ereg_replace(148, "&#034;", $this->$property); // right double quote
-                  $this->$property = ereg_replace(8226, "&#149;", $this->$property); // bullet
-                  $this->$property = ereg_replace(149, "&#149;", $this->$property); // bullet
-                  $this->$property = ereg_replace(8211, "&#150;", $this->$property); // en dash
-                  $this->$property = ereg_replace(150, "&#150;", $this->$property); // en dash
-                  $this->$property = ereg_replace(8212, "&#151;", $this->$property); // em dash
-                  $this->$property = ereg_replace(151, "&#151;", $this->$property); // em dash
-                  $this->$property = ereg_replace(8482, "&#153;", $this->$property); // trademark
-                  $this->$property = ereg_replace(153, "&#153;", $this->$property); // trademark
-                  $this->$property = ereg_replace(169, "&copy;", $this->$property); // copyright mark
-                  $this->$property = ereg_replace(174, "&reg;", $this->$property); // registration mark
-                */
-
-                $this->$property = ereg_replace(133, "", $this->$property); // ellipses
-                $this->$property = ereg_replace(8226, "", $this->$property); // double prime
-                $this->$property = ereg_replace(8216, "", $this->$property); // left single quote
-                $this->$property = ereg_replace(145, "", $this->$property); // left single quote
-                $this->$property = ereg_replace(8217, "", $this->$property); // right single quote
-                $this->$property = ereg_replace(146, "", $this->$property); // right single quote
-                $this->$property = ereg_replace(8220, "", $this->$property); // left double quote
-                $this->$property = ereg_replace(147, "", $this->$property); // left double quote
-                $this->$property = ereg_replace(8221, "", $this->$property); // right double quote
-                $this->$property = ereg_replace(148, "", $this->$property); // right double quote
-                $this->$property = ereg_replace(8226, "", $this->$property); // bullet
-                $this->$property = ereg_replace(149, "", $this->$property); // bullet
-                $this->$property = ereg_replace(8211, "", $this->$property); // en dash
-                $this->$property = ereg_replace(150, "", $this->$property); // en dash
-                $this->$property = ereg_replace(8212, "", $this->$property); // em dash
-                $this->$property = ereg_replace(151, "", $this->$property); // em dash
-                $this->$property = ereg_replace(8482, "", $this->$property); // trademark
-                $this->$property = ereg_replace(153, "", $this->$property); // trademark
-                $this->$property = ereg_replace(169, "", $this->$property); // copyright mark
-                $this->$property = ereg_replace(174, "", $this->$property); // registration mark
-            }
-        }
-    }
-
-    public function delete($table = null)
-    {
-        if ($table == null) {
-            $c = get_class($this);
-            $table = $c::TABLE;
-        }
-        $sql = 'DELETE FROM ' . $table . ' WHERE id = ' . $this->id;
-        $error = 'Could not delete row.';
+        $table = $this::TABLE;
+        $id = $this->id;
+        $sql = "DELETE FROM $table WHERE id = $id";
+        $error = "Could not delete row.";
         $this->query($sql, $error);
     }
 
@@ -932,13 +627,6 @@ class Entity
         return $this->val->valType($type, $field, $data, $optional, $message);
     }
 
-
-    /**
-     * get entity properties
-     *
-     *
-     *
-     */
     public function getProperties()
     {
         return $this->properties;